+++ crates/db-user/src/session_participants_migration_1.sql
@@ -1,0 +1,1 @@
+ALTER TABLE session_participants ADD COLUMN deleted BOOLEAN DEFAULT FALSE;
